+++ demo-app/package.json
@@ -18,7 +18,8 @@
     "grommet-theme-hpe": "^5.0.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^6.13.0"
+    "react-router-dom": "^6.13.0",
+    "styled-components": "^5.3.10"
   },
   "devDependencies": {
     "@vitejs/plugin-react": "^3.1.0",
